+++ lib/mailboxer/models/messageable.rb
@@ -59,9 +59,10 @@ module Mailboxer
       #as originator
       def send_message(recipients, msg_body, subject, sanitize_text=true, attachments=nil, message_timestamp = Time.now, with_email: true)
         convo = Mailboxer::ConversationBuilder.new({
-          :subject    => subject,
-          :created_at => message_timestamp,
-          :updated_at => message_timestamp
+          :subject       => subject,
+          :created_at    => message_timestamp,
+          :updated_at    => message_timestamp,
+          :sanitize_text => sanitize_text
         }).build
 
         message = Mailboxer::MessageBuilder.new({